Performance and functionality improvements
