+++ .gitignore
@@ -169,6 +169,8 @@ examples/permz_*,1,*.png
 examples/poro_*,1,*.png
 examples/porv_*,1,*.png
 examples/satnum_*,1,*.png
+examples/sgas_*,1,*.png
+examples/fgip.png
 tests/generic_deck
 
 # Python environment
