+++ MANIFEST.in
@@ -1,3 +1,4 @@
 include README.md
 include LICENSE
 include src/plopm/core/*
+include src/plopm/utils/*
